perbaikan fungsi edit foto menurut jk

+++ app/Controllers/User.php
@@ -442,7 +442,7 @@ class User extends BaseController
 			$avatar = $this->request->getFile('file_upload');
 			$avatar->move(ROOTPATH . '/public/img/components/user/userid_' . session('id_user'));
 
-			if ($foto != 'default.svg') {
+			if ($foto != $query1->jenis_kelamin.'/default.svg' && $foto != 'default.svg') {
 				$url = ROOTPATH . '/public/img/' . $foto;
 				if (is_file($url))
 					unlink($url);
@@ -472,14 +472,14 @@ class User extends BaseController
 		$query1 = $model->bukaProfile(session('id_alumni'))->getRow();
 		$foto = $query1->foto_profil;
 
-		if ($foto != 'default.svg') {
+		if ($foto != $query1->jenis_kelamin.'/default.svg' && $foto != 'default.svg') {
 			$url = ROOTPATH . '/public/img/' . $foto;
 			if (is_file($url))
 				unlink($url);
 		}
 
 		$data = [
-			'foto_profil' => 'default.svg'
+			'foto_profil' => $query1->jenis_kelamin.'/default.svg'
 		];
